Fix bug in interpolating complex frequency series

+++ gwpy/frequencyseries/frequencyseries.py
@@ -292,7 +292,8 @@ class FrequencySeries(Series):
         """
         f0 = self.f0.decompose().value
         N = (self.size - 1) * (self.df.decompose().value / df) + 1
-        fsamples = numpy.arange(0, numpy.rint(N), dtype=self.dtype) * df + f0
+        fsamples = numpy.arange(0, numpy.rint(N),
+                                dtype=self.real.dtype) * df + f0
         out = type(self)(numpy.interp(fsamples, self.frequencies.value,
                                       self.value))
         out.__array_finalize__(self)
